fix(Create entity with dependency graph)
* Close #17
- modification of the entities fields types to match the java version
- fix the correspondance between the label field of the frontend label entity and the name
field of the backend label entity
- fix the creation of an instance of an entity that references another
- addition of tests regarding the last point

* Removing tests that are no longer needed and fix the formatting.

+++ src/JHipsterNetSampleApplication/Models/OperationLabel.cs
@@ -1,9 +1,9 @@
 namespace JHipsterNetSampleApplication.Models {
     public class OperationLabel {
-        public int OperationId { get; set; }
+        public long OperationId { get; set; }
         public Operation Operation { get; set; }
 
-        public int LabelId { get; set; }
+        public long LabelId { get; set; }
         public Label Label { get; set; }
     }
 }
